pyproject.toml
[project]
name = "ai-agents-with-langchain-and-langgraph"
version = "0.1.0"
description = "ai-agents-with-langchain-and-langgraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.18",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
